requirements.txt
pytest~=8.0.2
sqlalchemy~=2.0.25
fastapi~=0.109.1
uvicorn~=0.25.0
redis~=5.1.0b4
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.7
pydantic~=2.5.3
h11~=0.14.0
pip~=24.0
attrs~=23.2.0
tornado~=6.4
Jinja2~=3.1.3
rsa~=4.9
pyasn1~=0.5.1
ecdsa~=0.18.0
typing_extensions~=4.9.0
sniffio~=1.3.0
six~=1.16.0
requests~=2.31.0
wcwidth~=0.2.13
pluggy~=1.4.0
iniconfig~=2.0.0
starlette~=0.36.3
pydantic_core~=2.14.6
websockets~=12.0
anyio~=4.2.0
click~=8.1.7
httptools~=0.6.1
watchfiles~=0.21.0
greenlet~=3.0.3
psutil~=5.9.7
decorator~=5.1.1
pycparser~=2.21
PyYAML~=6.0.1
psycopg2~=2.9.9
bcrypt~=4.1.2